docker-run: use /etc/hosts file from host
Its common in development environment that we have a lot of hosts
defined on /etc/hosts, so when you try to use some commands that use
docker container, maybe it dont work because its container can't find
hosts.

This commit solve this problem, mounting a volume that set /etc/hosts as
container host list.